LayoutTests/editing/deleting/merge-list-items-in-same-list.html

<!DOCTYPE html>
<html>
<body>
<div>
<ul contenteditable><li>one</li><li></li><li id="li">three</li></ul>
</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>
var li = document.getElementById('li');
var selection = window.getSelection();
selection.collapse(li, 0);
document.execCommand('Delete');
Markup.dump(document.querySelector('div'));
</script>
</body>
</html>
