LayoutTests/editing/deleting/merge-list-items-in-same-list-expected.txt

| "\n"
| <ul>
|   contenteditable=""
|   <li>
|     "one"
|   <li>
|     "<#selection-caret>three"
|     <br>
| "\n"
